Fix bug: missing kickoff platform
The saved start states didn't include the location of the kickoff
platform. So, when replayed, the platform would be missing, as if the
ground under the stance foot disappears as soon as the swing foot hits
the ground. This probably made the dynamics hard to learn.

Also, this commit removes some misplaced code from walker.py.