Massive refactor: Support 2D targets and use absolute coordinates
It was making my head hurt to remember which function parameters were
supposed to be in relative coordinates and which in absolute (and what
the coordinates were supposed to be relative to). This commit refactors
so everything is in absolute coordinates. The method `center_state` in
inverse_dynamics now handles moving things to relative coordinates for
the purposes of training and using the model.

This commit also changes targets to (x,y) instead of just (x), in
anticipation of future work where the targets will not all be in the y=0
line.

There are a number of other smaller changes, such as having Simbicon
objects load and serialize their own state, and cleaning up some log
messages.