Cleanly factor out State object; fix crash detection bugs
This is fairly major refactor of simbicon.py and walker.py.

1. current_observation() is now a structured object (a wrapper
around the underlying numpy array).

2. The state_complete / change_state interface between walker.py and
simbicon.py has been dramatically simplified. It is now a single function
(step_complete) with a single boolean return value.

3. In the process, this fixed some bugs. Now we can run step_learner
with small platforms, and even though it will crash a lot in the
learning process, it successfully learns a good step.