Try to better abstract model details from algorithm
This introduces a pair of functions `to_features`, `from_features`
that permute and switch signs for the entries of robot_skeleton.x
as appropriate. By wrapping low-level interactions with the model with
these functions, we can protect the rest of the code base from weird
bugs and cluttered code that stem from differences between models.

Note: there might be further bugs due to noncommutativity of Euler
angles. I'm not sure whether any of the models so far use a different
order of rotations.

It also removes visualization of the pelvis, because that was making it
hard to track the the robot's (x,y,z) location with put_dot.