Reorganize Simbicon params (FSMstate) object
Many params are valid for the entire step, not just for the UP or the
DOWN phase. This makes it so there is a single number that controls
those params for the entire step.

It does slightly change the behavior of the zero action. The
"adjust_up_targets" method now affects some numbers in the DOWN phase.
However, base behavior still seems reasonable. (I just need to rename
that method....)