Make step target a parameter
The Simbicon controller used self.target when (1) adjusting certain
parameters, such as swing hip angle, at the beginning of the step, (2)
deciding when to start the DOWN phase, and (3) calculating inverse
kinematics for the DOWN phase.

This commit makes the "target" used for these calculations a parameter,
although (1) this parameter is not "controllable" yet, and (2) it keeps
the IK_GAIN parameter, which is now a bit redundant except that the
basic controller probably won't work without it.