Allow named dots
This is a much more intuitive way to refer to dots than having to keep
track of how many dots have already been created.

Also fix a lingering bug in inverse_kinematics that I discovered while
doing this refactor. I also discovered that ik.test() is behaving oddly
and in fact appears to be broken, but I'm not going to fix it right now.