Wow I had a lot of bugs
This commit updates a lot of code in simbicon.py so that it will work
when traveling in arbitrary directions. Kind of crazy how even after all
of the refactors, so much of the code assumed positive X direction.

- Introduce an instance variable `self.target_direction`, updated at the
beginning of each step, to track the "direction" of the step.
- Measure distance from / velocity toward the target by projecting it to
that line.
- Also update crash detection so unit_normal (hopefully) works properly.

2D Simbicon still works. I updated 3D Simbicon testing code to allow
going in different directions. It kind of works if the angle is small,
but there are still clearly bugs I haven't found.