Fix bugs mentioned in previous commit
There is now no gimbal lock (unless the robot is lying on the ground,
which it won't ever do in normal operation).

The "yaw" angle is now more of a "heading" angle in that it is always
applied with respect to the global Z axis, not the head-to-toe axis of
the robot. This way, it will never affect the dynamics/direction of
gravity.

How I fixed it: There's probably a cleaner way to do this. I couldn't
find .skel file format documentation so I just tried random stuff until
it worked. I ended up breaking the previous "euler" joint (3 DOFs) into
a "universal" joint (2 DOFs) and a "revolute" joint (1 DOF). This
involved adding a new pelvis_aux2 bodynode. The axis
xyz settings don't seem to behave intuitively, so again I just tried
random settings until it worked.