Try some PD control for robot heading using stance hip
It kind of works, unless there's not enough friction and the stance leg
twists (instead of turning the rest of the body like it's supposed to).

This involved changing the axis_order for the hip joints so that "yaw"
behaves more like "heading". (I tried doing the same thing I did for the
root Euler angles in the commit 552bf62, but I couldn't figure out how
to modify the skel file properly.)

Also: in order to get it to work, you also need to *set* the
target_direction and target_heading in simbicon.py to match the heading
used when generating the current target. The current system of guessing
the right heading based on the line from the swing heel to the target is
a little too unstable.