Add method to calculate which way the stance foot is pointing
This factors out the target_heading calculation from simbicon.py so it
can also be used in inverse_kinematics.py. This also implements logic to
adjust that heading by multiples of 2*pi until it is "close" to some
reference heading.

The idea is to try using the new ik.heading() method to control yaw in
simbicon_3D (which may be more stable than using ROOT_YAW only), but I
haven't tested this yet.