Add test for inverse-Euler (doesn't pass yet)
As well as some other cleanup in inverse_kinematics.py.

The point of the test is to be able to take some starting state,
randomly perturb the joint angles, then change just the first six DOFs
so that the *left thigh* is in the same orientation as it was before.
I'm not quite sure how to do this, honesly.

I think this is a necessary first step to the follow-up test, which is,
given a random starting state and a target pelvis orientation, change
the left hip angles so that, if the thigh orientation stays the same,
then the pelvis will have the target orientation.