If USE_VIRTUAL_TORQUE is False, use new IK code to control heading
The new IK code doesn't seem to work; the robot slowly falls over. Maybe
I need to increase the gains. But for now, I want to commit this code
and make sure that 2D Simbicon is still working.

This also includes code to set the target heading. Ideally the
controller should be able to figure out a good target heading using just
the location of the target relative to the robot, but that doesn't work
yet. For now I need to set the heading manually.