step_learner can now learn basic step in 3D
To make this work, I did a few things:

1. Added a "toe-off" state to the FSM. This allows more dramatic use of
the swing ankle at the beginning of the step while still ending the step
with a heel strike.

2. Tinkered with the controllable parameters in 3D. It turns out three
parameters is enough (heading, ik_gain, and toe-off strength), but I had
to rescale the toe-off strength in order for random search to find the
proper parameter setting. (The 'manual_search' method is for manually
trying different parameter settings while tinkering/debugging.)

3. Fixed a bug in score calculation due to flipping the robot state but
not the target Z coordinate.

4. Decreased the learning rate. (2D learning still works fine at this
rate.)