Various refactors/minor fixes in pursuit of inverse_dynamics
1. Actions are now low-dimensional, rather than sparse. This is mostly
to make it easier to just print the whole list of parameters of the
model.

2. Remove RANSAC and training set perturbations (for now).

3. Only use 8 resets when collecting starting states (down from 16 to
save time).

4. Remove action retargeting. It would be too expensive to add three
action dimensions . . . and anyway it wasn't working and was hard to
understand.

5. Store the swing platform in extract_features, and don't filter to
observable features when storing training set. (Instead, filter to
observable features when training, after loading the data to memory.
It's nice to have all data available for debugging.)

6. Handle Z-flipping for actions in Simbicon3D rather than
InverseDynamics.

7. Rewrite test_train_state_storage to focus more on seeing how well the
model can fit the training samples. (Answer: very poorly...why?)