Refactor: use env.reset() to control render settings
Instead of env.simulate(). Mostly this is so we can suppress the "Seed:"
output when we're not rendering anything. It also makes it easier to
automatically decrease the render rate when we're recording video.
- Also: when using Runner, now you must call reset() before each run().

This also makes a number of changes to get the full ML pipeline running
in 3D, as well as some small changes to improve its 3D performance
slightly. It's still mostly garbage.

Many of these changes affect 2D as well, hopefully not adversely. One
change (increasing dist_mean to 0.47) is specifically for 2D, attempting
to fix the problem where, during evaluation, the robot accidentally
steps on two stairs at once. (Usually this would cause it to fall over.)