Properly implement dataset aggregation
I was suffering from the covariate shift problem because start_states
were taken from the end_states of the *expert*, not the learner (i.e.
the linear regressor). This commit majorly reorganizes the learning
algorithm. Now, the start states used in the training dataset come from
evaluate() rollouts that used the actual learner policy.

One nice side effect is that the initial start state collection can be
treated as an easy first stage of the curriculum where the step length
is not random.

To help keep track of the settings used at various stages of the
curriculum, I've added a "history" object as well.

Also: Split inverse_dynamics tests into their own file, and make it
easier to close video recorders.