Clean up interface for reducing problem dimensionality
`observable_features` and `controllable_params` are now set in
curriculum.py, rather than a hodgepodge of other locations.

- reinforcement.py should now be understandable and usable by people
other than just me.

- The observable features are now defined in terms of SWING_IDX and
STANCE_IDX rather than RIGHT_IDX and LEFT_IDX. I haven't quite found a
clean way to do this, but as long as filtering to the observable
features is always done on the return value of state.extract_features,
it should be fine.

- Some bugfixes and cleanup from the previous refactors.