Use Euler angles for Darwin root joint
The parametrization used by FreeJoint is better in that it doesn't have
the gimbal lock problem, but worse in that there seems to be no API to
transform between it and the Euler angles used by Simbicon.

Unfortunately there is also no clean API to set the axis order to YZX or
something that avoids gimbal lock when Y is the gravity axis. This means
I will now have to embark upon a wild refactor of all my code to make Z
the gravity axis.