Implement virtual torque via a target stance hip angle
Rather than manually overwriting that element of the torque vector. This
is to conform with the interface for the controller for the physical
Darwin robot. Mathematically, in the simulator, the torque should work
out to be the same (except that the target angle is set only at the
control frequency).

I'm not sure that this would *actually* work on the physical robot, but
hopefully it's closer.