Remove automatic hip joint targets and set some hip joint limits
The stance hip in particular tried to keep the torso upright by using
information about absolute pose, which we won't have on the hardware.

Then we need the joint limits to ensure that the learned gaits that
*don't* use that automatic targeting will remain feasible.

Also tweak the learn_llc target again a bit---for some reason it seems
to be better not to reward the algorithm directly for distance traveled,
but instead for number of steps taken.

I am not sure how well learn_llc performs with these changes; learning
is still in progress.