Add new baseline based on running CMA with an average of step rewards
It doesn't work very well, and I'm not quite sure why I spent so much
time on this. But the idea is that this uses the rewards at each step
rather than just the reward at the end of the episode.

I should probably also write a CMA baseline that uses entire episodes
and see how that fares. (It should at least be able to learn a little
bit, since that's what ARS can do.)