Add a new controllable param to 3D basic and change 3D curriculum
It was having a hard time controlling lateral placement of the swing
foot. The curriculum change modifies 3D according to what we did for 2D:
start on stepping stones right away, but with an easier offset
distribution. (This makes it less likely that the learner will learn
strange strategies that only work on large flat surfaces.)

This commit also modifes the step_learner tests a bit, to make them
better match what is actually encountered in the experiments.