Move the "early termination" code into evaluation settings
This way it won't, for example, interfere with training low-level gaits
for the Darwin simulator.

It also has the benefit of simplifying the experimental setup to bring
it closer to pure DAgger. Now, the evaluation environment should have
exactly the same distribution as the environment used to collect
training data (at least once we reach that stage of the curriculum).