Use full state when resetting environment to find expert labels
It turns out that, for the simple 3D model, the waist heading and pitch
DOFs are not included in current_observation(). Erasing this information
can make a significant difference in the final score for the step
(equivalent to ~3cm of accuracy in foot placement). I think being more
consistent about this could help with generalization error. Since we're
shamelessly using the simulator anyway, I now extract the raw DOFs using
world.robot_skeleton.x and use that when resetting the environment.

Similarly, rotating the heading angle when standardizing states loses
information (namely the target_heading being used by the Simbicon
controller, relative to the current heading of the robot, and therefore
also the heading angle of the stepping stones, etc.). So this removes
that rotation code.