Revert to code state used for test_cim_3D_9
This is the best-performing controller so far, but it depends on certain
code features that aren't captured in `history` (for example, it
requires a constant `target_heading` of zero). This commit memorializes
those settings (assuming I remember them correctly). I tested the
controller with these settings and it seems to perform comparably to
what it did during training, which indicates the settings are correct.