Fix bug when step ends not on a control tick
This was causing issues with reproducibility. If a step ends when the
simulation frame is e.g. 1523 (some non-round number), then when we reset
the environment and try to execute the step again, the behavior will be
different even if the initial states of all of the skeletons are the
same.

TODO: make it impossible to end a step not on a control tick.